+++ deepxube/base/heuristic.py
@@ -4,7 +4,7 @@ from typing import List, Any, TypeVar, Generic, cast, Tuple, Optional, Type, Pro
 import numpy as np
 from numpy.typing import NDArray
 
-from deepxube.base.domain import Domain, State, Goal, Action
+from deepxube.base.domain import State, Goal, Action
 from deepxube.base.nnet_input import NNetInput, PolicyNNetIn
 from deepxube.nnet.nnet_utils import NNetParInfo, nnet_batched, NNetPar, get_nnet_par_out
 from deepxube.utils import misc_utils
@@ -13,8 +13,6 @@ import torch
 from torch import nn, Tensor
 import torch.optim as optim
 from torch.optim.optimizer import Optimizer
-
-import random
 
 
 In = TypeVar('In', bound=NNetInput)
@@ -71,12 +69,16 @@ PNNetIn = TypeVar('PNNetIn', bound=PolicyNNetIn)
 
 
 class PolicyNNet(DeepXubeNNet[PNNetIn], ABC):
+    def __init__(self, nnet_input: PNNetIn, num_samp: int, **kwargs: Any):
+        super().__init__(nnet_input)
+        self.num_samp: int = num_samp
+
     @abstractmethod
     def forward(self, states_goals: List[Tensor]) -> List[Tensor]:
-        """ Condition on states and goals to sample actions
+        """ Condition on states and goals to sample self.num_samp actions
 
         :param states_goals:
-        :return: List of tensors representing actions with the last Tensor representing the probability density of actions
+        :return: List of tensors representing actions with the last Tensor representing the probability density of actions (N x num_samp x ...)
         """
         pass
 
@@ -95,13 +97,14 @@ def _flatten_list(data_l: List[Tensor]) -> Tensor:
 
 
 class PolicyVAE(PolicyNNet[PNNetIn]):
-    def __init__(self, nnet_input: PNNetIn, kl_weight: float, **kwargs: Any):
-        super().__init__(nnet_input)
+    def __init__(self, nnet_input: PNNetIn, num_samp: int, kl_weight: float, **kwargs: Any):
+        super().__init__(nnet_input, num_samp)
         self.norm_dist = torch.distributions.Normal(0, 1)
         self.kl_weight: float = kl_weight
         self.mse_criterion = nn.MSELoss()
 
     def forward(self, states_goals: List[Tensor]) -> List[Tensor]:
+        # TODO use num_samp
         z: Tensor = self.norm_dist.sample((states_goals[0].shape[0],) + self.latent_shape()).to(states_goals[0].device)
         recons: List[Tensor] = self.decode(states_goals, z)
         return recons + [self.norm_dist.log_prob(z).sum(dim=1, keepdim=True)]
@@ -176,7 +179,7 @@ HeurFn = Union[HeurFnV, HeurFnQ]
 
 @runtime_checkable
 class PolicyFn(Protocol):
-    def __call__(self, domain: Domain, states: List[State], goals: List[Goal]) -> Tuple[List[List[Action]], List[List[float]]]:
+    def __call__(self, states: List[State], goals: List[Goal]) -> Tuple[List[List[Action]], List[List[float]]]:
         """ Map states and goals to sampled actions along with their probability (or log probability) densities
 
         """
@@ -353,75 +356,24 @@ class HeurNNetParQIn(HeurNNetParQ, ABC):
         return q_vals_l
 
 
-def policy_fn_rand(domain: Domain, states: List[State], num_rand: int) -> Tuple[List[List[Action]], List[List[float]]]:
-    if num_rand == 0:
-        return [[] for _ in states], [[] for _ in states]
-
-    states_rep: List[List[State]] = []
-    for state in states:
-        states_rep.append([state] * num_rand)
-
-    states_rep_flat, split_idxs = misc_utils.flatten(states_rep)
-
-    actions_samp_flat: List[Action] = domain.sample_state_action(states_rep_flat)
-    actions_samp_l: List[List[Action]] = misc_utils.unflatten(actions_samp_flat, split_idxs)
-
-    probs_l: List[List[float]] = []
-    for actions_samp_i in actions_samp_l:
-        probs_l.append([1.0 / len(actions_samp_i)] * len(actions_samp_i))
-
-    return actions_samp_l, probs_l
-
-
-def _combine_nnet_with_rand(domain: Domain, actions_l: List[List[Action]], pdfs_l: List[List[float]], states: List[State],
-                            num_rand: int) -> Tuple[List[List[Action]], List[List[float]]]:
-    actions_rand_l: List[List[Action]] = policy_fn_rand(domain, states, num_rand)[0]
-
-    actions_comb_l: List[List[Action]] = []
-    pdfs_comb_l: List[List[float]] = []
-    # get nnet actions
-    for state_idx in range(len(states)):
-        actions_rand_i: List[Action] = actions_rand_l[state_idx]
-        actions_comb_l.append(actions_l[state_idx] + actions_rand_i)
-        pdfs_i: List[float] = pdfs_l[state_idx]
-        pdfs_comb_l.append(pdfs_i + random.choices(pdfs_i, k=len(actions_rand_i)))
-
-    return actions_comb_l, pdfs_comb_l
-
-
 class PolicyNNetPar(NNetPar[PolicyFn]):
-    def __init__(self, num_samp: int, num_rand: int):
-        self.num_samp: int = num_samp
-        self.num_rand: int = num_rand
-
     def get_nnet_fn(self, nnet: nn.Module, batch_size: Optional[int], device: torch.device, update_num: Optional[int]) -> PolicyFn:
         nnet.eval()
-        if (update_num is not None) and (update_num == 0):
-            def policy_fn(domain: Domain, states: List[State], goals: List[Goal]) -> Tuple[List[List[Action]], List[List[float]]]:
-                assert len(states) == len(goals)  # to stop PyCharm from complaining
-                return policy_fn_rand(domain, states, self.num_samp + self.num_rand)
-        else:
-            def policy_fn(domain: Domain, states: List[State], goals: List[Goal]) -> Tuple[List[List[Action]], List[List[float]]]:
-                inputs_nnet_rep: List[NDArray] = self._get_nnet_inputs_rep(states, goals, self.num_samp)
-                nnet_out_np: List[NDArray[np.float64]] = nnet_batched(nnet, inputs_nnet_rep, batch_size, device)
 
-                actions_l, pdfs_l = self._np_to_acts_and_pdfs(nnet_out_np[0:-1], nnet_out_np[-1], len(states), self.num_samp)
-                return _combine_nnet_with_rand(domain, actions_l, pdfs_l, states, self.num_rand)
+        def policy_fn(states: List[State], goals: List[Goal]) -> Tuple[List[List[Action]], List[List[float]]]:
+            inputs_nnet: List[NDArray] = self.to_np_fn(states, goals)
+            nnet_out_np: List[NDArray[np.float64]] = nnet_batched(nnet, inputs_nnet, batch_size, device)
+
+            return self._np_to_acts_and_pdfs(nnet_out_np[0:-1], nnet_out_np[-1], len(states))
 
         return policy_fn
 
     def get_nnet_par_fn(self, nnet_par_info: NNetParInfo, update_num: Optional[int]) -> PolicyFn:
-        if (update_num is not None) and (update_num == 0):
-            def policy_fn(domain: Domain, states: List[State], goals: List[Goal]) -> Tuple[List[List[Action]], List[List[float]]]:
-                assert len(states) == len(goals)  # to stop PyCharm from complaining
-                return policy_fn_rand(domain, states, self.num_samp + self.num_rand)
-        else:
-            def policy_fn(domain: Domain, states: List[State], goals: List[Goal]) -> Tuple[List[List[Action]], List[List[float]]]:
-                inputs_nnet_rep: List[NDArray] = self._get_nnet_inputs_rep(states, goals, self.num_samp)
-                nnet_out_np: List[NDArray[np.float64]] = get_nnet_par_out(inputs_nnet_rep, nnet_par_info)
+        def policy_fn(states: List[State], goals: List[Goal]) -> Tuple[List[List[Action]], List[List[float]]]:
+            inputs_nnet: List[NDArray] = self.to_np_fn(states, goals)
+            nnet_out_np: List[NDArray[np.float64]] = get_nnet_par_out(inputs_nnet, nnet_par_info)
 
-                actions_l, pdfs_l = self._np_to_acts_and_pdfs(nnet_out_np[0:-1], nnet_out_np[-1], len(states), self.num_samp)
-                return _combine_nnet_with_rand(domain, actions_l, pdfs_l, states, self.num_rand)
+            return self._np_to_acts_and_pdfs(nnet_out_np[0:-1], nnet_out_np[-1], len(states))
 
         return policy_fn
 
@@ -441,25 +393,22 @@ class PolicyNNetPar(NNetPar[PolicyFn]):
     def _nnet_out_to_actions(self, nnet_out: List[NDArray[np.float64]]) -> List[Action]:
         pass
 
-    def _get_nnet_inputs_rep(self, states: List[State], goals: List[Goal], num_samp: int) -> List[NDArray]:
-        inputs_nnet: List[NDArray] = self.to_np_fn(states, goals)
-        inputs_nnet_rep_interleave: List[NDArray] = []
-        for inputs_nnet_i in inputs_nnet:
-            inputs_nnet_rep_interleave.append(np.repeat(inputs_nnet_i, num_samp, axis=0))
-        return inputs_nnet_rep_interleave
-
-    def _np_to_acts_and_pdfs(self, actions_np: List[NDArray[np.float64]], pdfs_np: NDArray[np.float64], num_states: int,
-                             num_samp: int) -> Tuple[List[List[Action]], List[List[float]]]:
+    def _np_to_acts_and_pdfs(self, actions_np: List[NDArray[np.float64]], pdfs_np: NDArray[np.float64],
+                             num_states: int) -> Tuple[List[List[Action]], List[List[float]]]:
+        # assert dimensions match
         assert len(pdfs_np.shape) == 2
-        assert pdfs_np.shape[1] == 1
+        assert pdfs_np.shape[0] == num_states
+        for actions_np_i in actions_np:
+            assert actions_np_i.shape[0] == num_states
+            assert actions_np_i.shape[0] == pdfs_np.shape[0]
+            assert actions_np_i.shape[1] == pdfs_np.shape[1]
 
+        # convert to action object rep
         actions_l: List[List[Action]] = []
         pdfs_l: List[List[float]] = []
         for state_idx in range(num_states):
-            start_idx: int = state_idx * num_samp
-            end_idx: int = start_idx + num_samp
-            actions_np_state: List[NDArray[np.float64]] = [actions_np_i[start_idx:end_idx] for actions_np_i in actions_np]
-            pdfs_state: List[float] = pdfs_np[start_idx:end_idx, 0].tolist()
+            actions_np_state: List[NDArray[np.float64]] = [actions_np_i[state_idx] for actions_np_i in actions_np]
+            pdfs_state: List[float] = pdfs_np[state_idx, :].tolist()
 
             actions_l.append(self._nnet_out_to_actions(actions_np_state))
             pdfs_l.append(pdfs_state)
@@ -467,4 +416,5 @@ class PolicyNNetPar(NNetPar[PolicyFn]):
         return actions_l, pdfs_l
 
     def __repr__(self) -> str:
-        return f"{self.get_nnet()}\n#Samp: {self.num_samp}, #Rand Samp: {self.num_rand}"
+        nnet: PolicyNNet = self.get_nnet()
+        return f"{nnet}\n#Samp: {nnet.num_samp}"